+++ config/forms/webhook/google_analytics.php
@@ -1,0 +1,52 @@
+<?php
+/**
+ * @var array $triggers
+ * @var App\Environment $environment
+ * @var App\Http\Router $router
+ */
+
+return [
+    'method' => 'post',
+
+    'groups' => [
+        [
+            'use_grid' => true,
+            'elements' => [
+
+                'name' => [
+                    'text',
+                    [
+                        'label' => __('Web Hook Name'),
+                        'description' => __(
+                            'Choose a name for this webhook that will help you distinguish it from others. This will only be shown on the administration page.'
+                        ),
+                        'required' => true,
+                        'form_group_class' => 'col-md-6',
+                    ],
+                ],
+
+                'tracking_id' => [
+                    'text',
+                    [
+                        'label' => __('GA Property Tracking ID'),
+                        'description' => __('The property ID used to track live listeners.'),
+                        'belongsTo' => 'config',
+                        'required' => true,
+                        'form_group_class' => 'col-md-6',
+                    ],
+                ],
+
+                'submit' => [
+                    'submit',
+                    [
+                        'type' => 'submit',
+                        'label' => __('Save Changes'),
+                        'class' => 'ui-button btn-lg btn-primary',
+                        'form_group_class' => 'col-sm-12',
+                    ],
+                ],
+
+            ],
+        ],
+    ],
+];